mega_test1_s3_bucket_alternate_solution2_end_to_end_python/requirements.txt
boto3==1.34.19
matplotlib==3.3.4
speedtest-cli==2.1.3
